requirement.txt
Flask
spacy
pdfminer.six